@@ -116,6 +116,10 @@ export class ViewLines extends ViewPart implements IVisibleLinesHost<ViewLine>, 
 	private _horizontalRevealRequest: HorizontalRevealRequest | null;
 	private readonly _lastRenderedData: LastRenderedData;
 
+	// Sticky Scroll
+	private _stickyScrollEnabled: boolean;
+	private _maxNumberStickyLines: number;
+
 	constructor(context: ViewContext, linesContent: FastDomNode<HTMLElement>) {
 		super(context);
 		this._linesContent = linesContent;
@@ -155,6 +159,10 @@ export class ViewLines extends ViewPart implements IVisibleLinesHost<ViewLine>, 
 		this._lastRenderedData = new LastRenderedData();
 
 		this._horizontalRevealRequest = null;
+
+		// sticky scroll widget
+		this._stickyScrollEnabled = options.get(EditorOption.experimental).stickyScroll.enabled;
+		this._maxNumberStickyLines = options.get(EditorOption.experimental).stickyScroll.maxLineCount;
 	}
 
 	public override dispose(): void {
@@ -196,6 +204,11 @@ export class ViewLines extends ViewPart implements IVisibleLinesHost<ViewLine>, 
 		this._cursorSurroundingLines = options.get(EditorOption.cursorSurroundingLines);
 		this._cursorSurroundingLinesStyle = options.get(EditorOption.cursorSurroundingLinesStyle);
 		this._canUseLayerHinting = !options.get(EditorOption.disableLayerHinting);
+
+		// sticky scroll
+		this._stickyScrollEnabled = options.get(EditorOption.experimental).stickyScroll.enabled;
+		this._maxNumberStickyLines = options.get(EditorOption.experimental).stickyScroll.maxLineCount;
+
 		applyFontInfo(this.domNode, fontInfo);
 
 		this._onOptionsMaybeChanged();
@@ -667,22 +680,30 @@ export class ViewLines extends ViewPart implements IVisibleLinesHost<ViewLine>, 
 
 		const shouldIgnoreScrollOff = (source === 'mouse' || minimalReveal) && this._cursorSurroundingLinesStyle === 'default';
 
+		let paddingTop: number = 0;
+		let paddingBottom: number = 0;
+
 		if (!shouldIgnoreScrollOff) {
 			const context = Math.min((viewportHeight / this._lineHeight) / 2, this._cursorSurroundingLines);
-			boxStartY -= context * this._lineHeight;
-			boxEndY += Math.max(0, (context - 1)) * this._lineHeight;
+			if (this._stickyScrollEnabled) {
+				paddingTop = Math.max(context, this._maxNumberStickyLines) * this._lineHeight;
+			} else {
+				paddingTop = context * this._lineHeight;
+			}
+			paddingBottom = Math.max(0, (context - 1)) * this._lineHeight;
 		} else {
 			if (!minimalReveal) {
 				// Reveal one more line above (this case is hit when dragging)
-				boxStartY -= this._lineHeight;
+				paddingTop = this._lineHeight;
 			}
 		}
-
 		if (verticalType === viewEvents.VerticalRevealType.Simple || verticalType === viewEvents.VerticalRevealType.Bottom) {
 			// Reveal one line more when the last line would be covered by the scrollbar - arrow down case or revealing a line explicitly at bottom
-			boxEndY += (minimalReveal ? this._horizontalScrollbarHeight : this._lineHeight);
+			paddingBottom += (minimalReveal ? this._horizontalScrollbarHeight : this._lineHeight);
 		}
 
+		boxStartY -= paddingTop;
+		boxEndY += paddingBottom;
 		let newScrollTop: number;
 
 		if (boxEndY - boxStartY > viewportHeight) {
