@@ -374,10 +374,27 @@ export class MergeConflictGutterItemView extends Disposable implements IGutterIt
 		const margin = checkboxHeight;
 
 		const effectiveCheckboxTop = top + middleHeight;
-		const clamped1 = clamp(effectiveCheckboxTop, margin, viewTop + viewHeight - margin - checkboxHeight);
-		const clamped2 = clamp(clamped1, top + margin, top + height - checkboxHeight - margin);
 
-		this.checkboxDiv.style.top = `${clamped2 - top}px`;
+		const preferredViewPortRange = [
+			margin,
+			viewTop + viewHeight - margin - checkboxHeight
+		];
+
+		const preferredParentRange = [
+			top + margin,
+			top + height - checkboxHeight - margin
+		];
+
+		const parentRange = [
+			top,
+			top + height - checkboxHeight
+		];
+
+		const clamped1 = clampIfIntervalIsNonEmpty(effectiveCheckboxTop, preferredViewPortRange[0], preferredViewPortRange[1]);
+		const clamped2 = clampIfIntervalIsNonEmpty(clamped1, preferredParentRange[0], preferredParentRange[1]);
+		const clamped3 = clamp(clamped2, parentRange[0], parentRange[1]);
+
+		this.checkboxDiv.style.top = `${clamped3 - top}px`;
 
 		this.target.classList.remove('multi-line');
 		this.target.classList.remove('single-line');
@@ -390,4 +407,11 @@ export class MergeConflictGutterItemView extends Disposable implements IGutterIt
 			this.item.set(baseRange, tx);
 		});
 	}
+}
+
+function clampIfIntervalIsNonEmpty(value: number, min: number, max: number): number {
+	if (min >= max) {
+		return value;
+	}
+	return Math.min(Math.max(value, min), max);
 }
